ISPN-2291: Moved the BaseRpcManager lock handling into the Replication and the Invalidation interceptors.
